Break out `{Cargo,}Miri` into their own `miri_tests` module
